requirements.txt
# ===== Core ML =====
tensorflow==2.15.0
keras==2.15.0
scikit-learn==1.4.0
xgboost==2.0.3
numpy==1.26.4
pandas==2.2.0
Pillow==10.2.0
matplotlib==3.8.2
seaborn==0.13.2
opencv-python-headless==4.9.0.80

# ===== Frontend =====
streamlit==1.31.0
plotly==5.18.0

# ===== Cloud SDKs =====
boto3==1.34.0
functions-framework==3.*

# ===== Utilities =====
requests==2.31.0
joblib==1.3.2
tqdm==4.66.1